OAK-1061: Add option for database name
git-svn-id: https://svn.apache.org/repos/asf/jackrabbit/oak/trunk@1528400 13f79535-47bb-0310-9956-ffa450edef68

+++ oak-run/src/main/java/org/apache/jackrabbit/oak/fixture/OakRepositoryFixture.java
@@ -79,8 +79,10 @@ public abstract class OakRepositoryFixture implements RepositoryFixture {
     }
 
     public static RepositoryFixture getMongo(
-            final String host, final int port, final long cacheSize) {
+            final String host, final int port, final String database,
+            final boolean dropDBAfterTest, final long cacheSize) {
         return new OakRepositoryFixture("Oak-Mongo") {
+            private String dbName = database != null ? database : unique;
             private MongoMK[] kernels;
             @Override
             public Repository[] setUpCluster(int n) throws Exception {
@@ -88,10 +90,10 @@ public abstract class OakRepositoryFixture implements RepositoryFixture {
                 kernels = new MongoMK[cluster.length];
                 for (int i = 0; i < cluster.length; i++) {
                     MongoConnection mongo =
-                            new MongoConnection(host, port, unique);
+                            new MongoConnection(host, port, dbName);
                     kernels[i] = new MongoMK.Builder().
                             setMongoDB(mongo.getDB()).
-                            setClusterId(i).open();
+                            setClusterId(i).setLogging(false).open();
                     Oak oak = new Oak(new KernelNodeStore(kernels[i], cacheSize));
                     cluster[i] = new Jcr(oak).createRepository();
                 }
@@ -102,13 +104,15 @@ public abstract class OakRepositoryFixture implements RepositoryFixture {
                 for (MongoMK kernel : kernels) {
                     kernel.dispose();
                 }
-                try {
-                    MongoConnection mongo =
-                            new MongoConnection(host, port, unique);
-                    mongo.getDB().dropDatabase();
-                    mongo.close();
-                } catch (Exception e) {
-                    throw new RuntimeException(e);
+                if (dropDBAfterTest) {
+                    try {
+                        MongoConnection mongo =
+                                new MongoConnection(host, port, dbName);
+                        mongo.getDB().dropDatabase();
+                        mongo.close();
+                    } catch (Exception e) {
+                        throw new RuntimeException(e);
+                    }
                 }
             }
         };
